Automatically tag rcran image with latest on Jenkins
We deploy the `docker-rstats` images to our worker. The `rstats` image uses this `docker-rcran` image as its base image.

No need to manually verify this image before tagging it with latest. The deployment to our worker is gated for the rstats image. Once we have better tests, we should automatically tag the `rstats` image with `latest` too on a green run.