R 4.4.0 (#79)
This is to update the image to use R 4.4.0,

we used an internal method from the util package to list dependencies,
seem like a refactor happened that cause the pipeline to break. the
refactor seem minimal (it was to speed up windows installs), but i
couldn't seem to decipher what datatype it wanted.

Easiest path forward is to integrate the old algo. into our image, and
not depend on an internal method. cleaned it up to make it readable.
https://b.corp.google.com/issues/282922446

passed build in commit,
[aa5cf28](https://github.com/Kaggle/docker-rcran/pull/79/commits/aa5cf285df9fac5b744151f45d63d813afd74bf3)

 I'm new to r, bear with me


#chavezcalderon-q2-fixit